move custom timings to new header file
